+++ history.md
@@ -1,7 +1,12 @@
 # 版本记录
+## 0.3.0 (2019-01-13)
+* [ADD] 新增 `print.buildTree()`
+* [ADD] 新增 `print.help()`
+
 ## 0.2.0 (2018-10-25)
 * [ADD] 新增 `print.log`, `print.fn`
 
 ## 0.1.0 (2018-10-18)
 * [ADD] 诞生
+* [ADD] 新增 `print.borderBox()`